+++ cruise-control/src/main/java/com/linkedin/kafka/cruisecontrol/servlet/KafkaCruiseControlServlet.java
@@ -11,6 +11,7 @@ import com.linkedin.kafka.cruisecontrol.servlet.parameters.AddedOrRemovedBrokerP
 import com.linkedin.kafka.cruisecontrol.servlet.parameters.BootstrapParameters;
 import com.linkedin.kafka.cruisecontrol.servlet.parameters.ClusterLoadParameters;
 import com.linkedin.kafka.cruisecontrol.servlet.parameters.DemoteBrokerParameters;
+import com.linkedin.kafka.cruisecontrol.servlet.parameters.GoalBasedOptimizationParameters;
 import com.linkedin.kafka.cruisecontrol.servlet.parameters.KafkaClusterStateParameters;
 import com.linkedin.kafka.cruisecontrol.servlet.parameters.PartitionLoadParameters;
 import com.linkedin.kafka.cruisecontrol.servlet.parameters.BaseParameters;
@@ -22,15 +23,14 @@ import com.linkedin.kafka.cruisecontrol.servlet.parameters.UserTasksParameters;
 import com.linkedin.kafka.cruisecontrol.async.AsyncKafkaCruiseControl;
 import com.linkedin.kafka.cruisecontrol.async.OperationFuture;
 import com.linkedin.kafka.cruisecontrol.config.KafkaCruiseControlConfig;
-import com.linkedin.kafka.cruisecontrol.monitor.ModelCompletenessRequirements;
 import com.linkedin.kafka.cruisecontrol.servlet.response.CruiseControlResponse;
-import com.linkedin.kafka.cruisecontrol.servlet.response.KafkaBootstrapResult;
-import com.linkedin.kafka.cruisecontrol.servlet.response.KafkaCruiseControlState;
-import com.linkedin.kafka.cruisecontrol.servlet.response.KafkaPauseSampling;
-import com.linkedin.kafka.cruisecontrol.servlet.response.KafkaResumeSampling;
-import com.linkedin.kafka.cruisecontrol.servlet.response.KafkaStopProposalExecution;
-import com.linkedin.kafka.cruisecontrol.servlet.response.KafkaTrainResult;
-import com.linkedin.kafka.cruisecontrol.servlet.response.KafkaUserTaskState;
+import com.linkedin.kafka.cruisecontrol.servlet.response.BootstrapResult;
+import com.linkedin.kafka.cruisecontrol.servlet.response.CruiseControlState;
+import com.linkedin.kafka.cruisecontrol.servlet.response.PauseSamplingResult;
+import com.linkedin.kafka.cruisecontrol.servlet.response.ResumeSamplingResult;
+import com.linkedin.kafka.cruisecontrol.servlet.response.StopProposalExecutionResult;
+import com.linkedin.kafka.cruisecontrol.servlet.response.TrainResult;
+import com.linkedin.kafka.cruisecontrol.servlet.response.UserTaskState;
 import java.io.IOException;
 import java.io.PrintWriter;
 import java.io.StringWriter;
@@ -56,6 +56,8 @@ import static com.linkedin.kafka.cruisecontrol.servlet.KafkaCruiseControlServlet
 import static com.linkedin.kafka.cruisecontrol.servlet.parameters.ParameterUtils.hasValidParameters;
 import static com.linkedin.kafka.cruisecontrol.servlet.parameters.ParameterUtils.wantJSON;
 import static com.linkedin.kafka.cruisecontrol.servlet.parameters.ParameterUtils.DataFrom;
+import static com.linkedin.kafka.cruisecontrol.servlet.response.ResponseUtils.returnProgress;
+import static com.linkedin.kafka.cruisecontrol.servlet.response.ResponseUtils.writeErrorResponse;
 import static javax.servlet.http.HttpServletResponse.SC_OK;
 import static javax.servlet.http.HttpServletResponse.SC_INTERNAL_SERVER_ERROR;
 
@@ -320,7 +322,7 @@ public class KafkaCruiseControlServlet extends HttpServlet {
     }
 
     _asyncKafkaCruiseControl.bootstrapLoadMonitor(parameters);
-    new KafkaBootstrapResult().writeSuccessResponse(parameters, response);
+    new BootstrapResult().writeSuccessResponse(parameters, response);
   }
 
   private void train(HttpServletRequest request, HttpServletResponse response) throws Exception {
@@ -331,7 +333,7 @@ public class KafkaCruiseControlServlet extends HttpServlet {
     }
 
     _asyncKafkaCruiseControl.trainLoadModel(parameters);
-    new KafkaTrainResult().writeSuccessResponse(parameters, response);
+    new TrainResult().writeSuccessResponse(parameters, response);
   }
 
   private boolean getClusterLoad(HttpServletRequest request, HttpServletResponse response) throws Exception {
@@ -381,7 +383,7 @@ public class KafkaCruiseControlServlet extends HttpServlet {
       return true;
     }
 
-    GoalsAndRequirements goalsAndRequirements =
+    GoalBasedOptimizationParameters.GoalsAndRequirements goalsAndRequirements =
         getGoalsAndRequirements(request, response, parameters.goals(), parameters.dataFrom(),
                                 parameters.ignoreProposalCache(), parameters.useReadyDefaultGoals());
     if (goalsAndRequirements == null) {
@@ -435,7 +437,7 @@ public class KafkaCruiseControlServlet extends HttpServlet {
       return true;
     }
 
-    GoalsAndRequirements goalsAndRequirements =
+    GoalBasedOptimizationParameters.GoalsAndRequirements goalsAndRequirements =
         getGoalsAndRequirements(request, response, parameters.goals(), parameters.dataFrom(), false, parameters.useReadyDefaultGoals());
     if (goalsAndRequirements == null) {
       return false;
@@ -472,7 +474,7 @@ public class KafkaCruiseControlServlet extends HttpServlet {
       return true;
     }
 
-    GoalsAndRequirements goalsAndRequirements =
+    GoalBasedOptimizationParameters.GoalsAndRequirements goalsAndRequirements =
         getGoalsAndRequirements(request, response, parameters.goals(), parameters.dataFrom(), false, parameters.useReadyDefaultGoals());
     if (goalsAndRequirements == null) {
       return false;
@@ -516,7 +518,7 @@ public class KafkaCruiseControlServlet extends HttpServlet {
       return;
     }
     _asyncKafkaCruiseControl.stopProposalExecution();
-    new KafkaStopProposalExecution().writeSuccessResponse(parameters, response);
+    new StopProposalExecutionResult().writeSuccessResponse(parameters, response);
   }
 
   private void pauseSampling(HttpServletRequest request, HttpServletResponse response) throws IOException {
@@ -526,7 +528,7 @@ public class KafkaCruiseControlServlet extends HttpServlet {
       return;
     }
     _asyncKafkaCruiseControl.pauseLoadMonitorActivity();
-    new KafkaPauseSampling().writeSuccessResponse(parameters, response);
+    new PauseSamplingResult().writeSuccessResponse(parameters, response);
   }
 
   private void resumeSampling(HttpServletRequest request, HttpServletResponse response) throws IOException {
@@ -536,7 +538,7 @@ public class KafkaCruiseControlServlet extends HttpServlet {
       return;
     }
     _asyncKafkaCruiseControl.resumeLoadMonitorActivity();
-    new KafkaResumeSampling().writeSuccessResponse(parameters, response);
+    new ResumeSamplingResult().writeSuccessResponse(parameters, response);
   }
 
   private CruiseControlResponse getAndMaybeReturnProgress(HttpServletRequest request,
@@ -555,29 +557,29 @@ public class KafkaCruiseControlServlet extends HttpServlet {
   }
 
   // package private for testing.
-  GoalsAndRequirements getGoalsAndRequirements(HttpServletRequest request,
-                                               HttpServletResponse response,
-                                               List<String> userProvidedGoals,
-                                               DataFrom dataFrom,
-                                               boolean ignoreCache,
-                                               boolean useReadyDefaultGoals) throws Exception {
+  GoalBasedOptimizationParameters.GoalsAndRequirements getGoalsAndRequirements(HttpServletRequest request,
+                                                                               HttpServletResponse response,
+                                                                               List<String> userProvidedGoals,
+                                                                               DataFrom dataFrom,
+                                                                               boolean ignoreCache,
+                                                                               boolean useReadyDefaultGoals) throws Exception {
     if (!userProvidedGoals.isEmpty() || dataFrom == DataFrom.VALID_PARTITIONS) {
-      return new GoalsAndRequirements(userProvidedGoals, getRequirements(dataFrom));
+      return new GoalBasedOptimizationParameters.GoalsAndRequirements(userProvidedGoals, getRequirements(dataFrom));
     }
 
     CruiseControlStateParameters parameters = new CruiseControlStateParameters(null);
-    parameters.setSubstates(new HashSet<>(Arrays.asList(KafkaCruiseControlState.SubState.ANALYZER,
-                                                        KafkaCruiseControlState.SubState.MONITOR)));
+    parameters.setSubstates(new HashSet<>(Arrays.asList(CruiseControlState.SubState.ANALYZER,
+                                                        CruiseControlState.SubState.MONITOR)));
 
     CruiseControlResponse state = getAndMaybeReturnProgress(request, response,
                                                             () -> _asyncKafkaCruiseControl.state(parameters, _userTaskManager));
     if (state == null) {
       return null;
     }
-    int availableWindows = ((KafkaCruiseControlState) state).monitorState().numValidWindows();
+    int availableWindows = ((CruiseControlState) state).monitorState().numValidWindows();
     List<String> allGoals = new ArrayList<>();
     List<String> readyGoals = new ArrayList<>();
-    ((KafkaCruiseControlState) state).analyzerState().readyGoals().forEach((goal, ready) -> {
+    ((CruiseControlState) state).analyzerState().readyGoals().forEach((goal, ready) -> {
       allGoals.add(goal.name());
       if (ready) {
         readyGoals.add(goal.name());
@@ -585,16 +587,16 @@ public class KafkaCruiseControlServlet extends HttpServlet {
     });
     if (allGoals.size() == readyGoals.size()) {
       // If all the goals work, use it.
-      return new GoalsAndRequirements(ignoreCache ? allGoals : Collections.emptyList(), null);
+      return new GoalBasedOptimizationParameters.GoalsAndRequirements(ignoreCache ? allGoals : Collections.emptyList(), null);
     } else if (availableWindows > 0) {
       // If some valid windows are available, use it.
-      return new GoalsAndRequirements(ignoreCache ? allGoals : Collections.emptyList(), getRequirements(dataFrom));
+      return new GoalBasedOptimizationParameters.GoalsAndRequirements(ignoreCache ? allGoals : Collections.emptyList(), getRequirements(dataFrom));
     } else if (useReadyDefaultGoals && readyGoals.size() > 0) {
       // If no window is valid but some goals are ready, use them if using ready goals is permitted.
-      return new GoalsAndRequirements(readyGoals, null);
+      return new GoalBasedOptimizationParameters.GoalsAndRequirements(readyGoals, null);
     } else {
       // Ok, use default setting and let it throw exception.
-      return new GoalsAndRequirements(Collections.emptyList(), null);
+      return new GoalBasedOptimizationParameters.GoalsAndRequirements(Collections.emptyList(), null);
     }
   }
 
@@ -605,26 +607,8 @@ public class KafkaCruiseControlServlet extends HttpServlet {
       return;
     }
 
-    KafkaUserTaskState kafkaUserTaskState = new KafkaUserTaskState(_userTaskManager.getActiveUserTasks(),
-                                                                   _userTaskManager.getCompletedUserTasks());
-    kafkaUserTaskState.writeSuccessResponse(parameters, response);
-  }
-
-  static class GoalsAndRequirements {
-    private final List<String> _goals;
-    private final ModelCompletenessRequirements _requirements;
-
-    private GoalsAndRequirements(List<String> goals, ModelCompletenessRequirements requirements) {
-      _goals = goals; // An empty list indicates the default goals.
-      _requirements = requirements;
-    }
-
-    List<String> goals() {
-      return _goals;
-    }
-
-    ModelCompletenessRequirements requirements() {
-      return _requirements;
-    }
+    UserTaskState userTaskState = new UserTaskState(_userTaskManager.getActiveUserTasks(),
+                                                    _userTaskManager.getCompletedUserTasks());
+    userTaskState.writeSuccessResponse(parameters, response);
   }
 }
