+++ cruise-control/src/main/java/com/linkedin/kafka/cruisecontrol/servlet/KafkaCruiseControlServletUtils.java
@@ -4,26 +4,20 @@
 
 package com.linkedin.kafka.cruisecontrol.servlet;
 
-import com.google.gson.Gson;
-import com.google.gson.GsonBuilder;
-import com.linkedin.kafka.cruisecontrol.async.OperationFuture;
 import com.linkedin.kafka.cruisecontrol.monitor.ModelCompletenessRequirements;
 import java.io.IOException;
-import java.io.OutputStream;
 import java.io.PrintWriter;
 import java.io.StringWriter;
 import java.io.UnsupportedEncodingException;
 import java.net.URLEncoder;
 import java.nio.charset.StandardCharsets;
-import java.util.HashMap;
-import java.util.Map;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
 import static com.linkedin.kafka.cruisecontrol.servlet.parameters.ParameterUtils.endPoint;
 import static com.linkedin.kafka.cruisecontrol.servlet.parameters.ParameterUtils.wantJSON;
 import static com.linkedin.kafka.cruisecontrol.servlet.parameters.ParameterUtils.DataFrom;
-import static javax.servlet.http.HttpServletResponse.SC_OK;
+import static com.linkedin.kafka.cruisecontrol.servlet.response.ResponseUtils.writeErrorResponse;
 import static javax.servlet.http.HttpServletResponse.SC_NOT_FOUND;
 import static javax.servlet.http.HttpServletResponse.SC_BAD_REQUEST;
 
@@ -32,12 +26,7 @@ import static javax.servlet.http.HttpServletResponse.SC_BAD_REQUEST;
  * The util class for Kafka Cruise Control servlet.
  */
 public class KafkaCruiseControlServletUtils {
-  public static final int JSON_VERSION = 1;
-  public static final String VERSION = "version";
   public static final String REQUEST_URI = "/KAFKACRUISECONTROL/";
-  private static final String STACK_TRACE = "stackTrace";
-  private static final String ERROR_MESSAGE = "errorMessage";
-  private static final String PROGRESS = "progress";
 
   private KafkaCruiseControlServletUtils() {
 
@@ -71,24 +60,6 @@ public class KafkaCruiseControlServletUtils {
     return s == null ? null : URLEncoder.encode(s, StandardCharsets.UTF_8.name());
   }
 
-  public static void setResponseCode(HttpServletResponse response, int code, boolean json) {
-    response.setStatus(code);
-    response.setContentType(json ? "application/json" : "text/plain");
-    response.setCharacterEncoding(StandardCharsets.UTF_8.name());
-    response.setHeader("Access-Control-Allow-Origin", "*");
-    response.setHeader("Access-Control-Request-Method", "OPTIONS, GET, POST");
-  }
-
-  private static void writeResponseToOutputStream(HttpServletResponse response, int responseCode, boolean json,
-                                                  String responseMsg)
-      throws IOException {
-    OutputStream out = response.getOutputStream();
-    setResponseCode(response, responseCode, json);
-    response.setContentLength(responseMsg.length());
-    out.write(responseMsg.getBytes(StandardCharsets.UTF_8));
-    out.flush();
-  }
-
   /**
    * Returns the endpoint if the request contains a valid one, otherwise (1) writes the error response to the given HTTP
    * response and (2) returns null.
@@ -117,40 +88,6 @@ public class KafkaCruiseControlServletUtils {
     StringWriter sw = new StringWriter();
     ure.printStackTrace(new PrintWriter(sw));
     writeErrorResponse(response, sw.toString(), errorMessage, SC_BAD_REQUEST, wantJSON(request));
-  }
-
-  public static void writeErrorResponse(HttpServletResponse response,
-                                        String stackTrace,
-                                        String errorMessage,
-                                        int responseCode,
-                                        boolean json)
-      throws IOException {
-    String responseMsg;
-    if (json) {
-      Map<String, Object> exceptionMap = new HashMap<>();
-      exceptionMap.put(VERSION, JSON_VERSION);
-      exceptionMap.put(STACK_TRACE, stackTrace);
-      exceptionMap.put(ERROR_MESSAGE, errorMessage);
-      Gson gson = new Gson();
-      responseMsg = gson.toJson(exceptionMap);
-    } else {
-      responseMsg = errorMessage == null ? "" : errorMessage;
-    }
-    writeResponseToOutputStream(response, responseCode, json, responseMsg);
-  }
-
-  static void returnProgress(HttpServletResponse response, OperationFuture future, boolean json) throws IOException {
-    String responseMsg;
-    if (json) {
-      Map<String, Object> respMap = new HashMap<>();
-      respMap.put(VERSION, JSON_VERSION);
-      respMap.put(PROGRESS, future.getJsonArray());
-      Gson gson = new GsonBuilder().serializeNulls().serializeSpecialFloatingPointValues().create();
-      responseMsg = gson.toJson(respMap);
-    } else {
-      responseMsg = future.progressString();
-    }
-    writeResponseToOutputStream(response, SC_OK, json, responseMsg);
   }
 
   static ModelCompletenessRequirements getRequirements(DataFrom dataFrom) {
