+++ cruise-control/src/main/java/com/linkedin/kafka/cruisecontrol/servlet/parameters/AddedOrRemovedBrokerParameters.java
@@ -6,21 +6,16 @@ package com.linkedin.kafka.cruisecontrol.servlet.parameters;
 
 import java.io.UnsupportedEncodingException;
 import java.util.List;
-import java.util.regex.Pattern;
 import javax.servlet.http.HttpServletRequest;
 
 
-public class AddedOrRemovedBrokerParameters extends KafkaOptimizationParameters {
+public class AddedOrRemovedBrokerParameters extends GoalBasedOptimizationParameters {
   private List<Integer> _brokerIds;
   private Integer _concurrentPartitionMovements;
   private Integer _concurrentLeaderMovements;
   private boolean _dryRun;
   private boolean _throttleAddedOrRemovedBrokers;
   private boolean _skipHardGoalCheck;
-  private ParameterUtils.DataFrom _dataFrom;
-  private boolean _useReadyDefaultGoals;
-  private List<String> _goals;
-  private Pattern _excludedTopics;
 
   public AddedOrRemovedBrokerParameters(HttpServletRequest request) {
     super(request);
@@ -35,10 +30,6 @@ public class AddedOrRemovedBrokerParameters extends KafkaOptimizationParameters 
     _concurrentPartitionMovements = ParameterUtils.concurrentMovements(_request, true);
     _concurrentLeaderMovements = ParameterUtils.concurrentMovements(_request, false);
     _skipHardGoalCheck = ParameterUtils.skipHardGoalCheck(_request);
-    _dataFrom = ParameterUtils.getDataFrom(_request);
-    _useReadyDefaultGoals = ParameterUtils.useReadyDefaultGoals(_request);
-    _goals = ParameterUtils.getGoals(_request);
-    _excludedTopics = ParameterUtils.excludedTopics(_request);
   }
 
   public List<Integer> brokerIds() {
@@ -63,21 +54,5 @@ public class AddedOrRemovedBrokerParameters extends KafkaOptimizationParameters 
 
   public boolean skipHardGoalCheck() {
     return _skipHardGoalCheck;
-  }
-
-  public ParameterUtils.DataFrom dataFrom() {
-    return _dataFrom;
-  }
-
-  public boolean useReadyDefaultGoals() {
-    return _useReadyDefaultGoals;
-  }
-
-  public List<String> goals() {
-    return _goals;
-  }
-
-  public Pattern excludedTopics() {
-    return _excludedTopics;
   }
 }
