+++ cruise-control/src/main/java/com/linkedin/kafka/cruisecontrol/servlet/parameters/RebalanceParameters.java
@@ -5,20 +5,17 @@
 package com.linkedin.kafka.cruisecontrol.servlet.parameters;
 
 import java.io.UnsupportedEncodingException;
-import java.util.List;
-import java.util.regex.Pattern;
 import javax.servlet.http.HttpServletRequest;
 
 
-public class RebalanceParameters extends KafkaOptimizationParameters {
+/**
+ * Parameters for {@link com.linkedin.kafka.cruisecontrol.servlet.EndPoint#REBALANCE}
+ */
+public class RebalanceParameters extends GoalBasedOptimizationParameters {
   private boolean _dryRun;
   private Integer _concurrentPartitionMovements;
   private Integer _concurrentLeaderMovements;
   private boolean _skipHardGoalCheck;
-  private ParameterUtils.DataFrom _dataFrom;
-  private boolean _useReadyDefaultGoals;
-  private List<String> _goals;
-  private Pattern _excludedTopics;
 
   public RebalanceParameters(HttpServletRequest request) {
     super(request);
@@ -31,10 +28,6 @@ public class RebalanceParameters extends KafkaOptimizationParameters {
     _concurrentPartitionMovements = ParameterUtils.concurrentMovements(_request, true);
     _concurrentLeaderMovements = ParameterUtils.concurrentMovements(_request, false);
     _skipHardGoalCheck = ParameterUtils.skipHardGoalCheck(_request);
-    _dataFrom = ParameterUtils.getDataFrom(_request);
-    _useReadyDefaultGoals = ParameterUtils.useReadyDefaultGoals(_request);
-    _goals = ParameterUtils.getGoals(_request);
-    _excludedTopics = ParameterUtils.excludedTopics(_request);
   }
 
   public boolean dryRun() {
@@ -51,21 +44,5 @@ public class RebalanceParameters extends KafkaOptimizationParameters {
 
   public boolean skipHardGoalCheck() {
     return _skipHardGoalCheck;
-  }
-
-  public ParameterUtils.DataFrom dataFrom() {
-    return _dataFrom;
-  }
-
-  public boolean useReadyDefaultGoals() {
-    return _useReadyDefaultGoals;
-  }
-
-  public List<String> goals() {
-    return _goals;
-  }
-
-  public Pattern excludedTopics() {
-    return _excludedTopics;
   }
 }
